# Introduce and use Context type