We introduce a new elasticapm.Context type. This type stores transaction/error context, providing
methods for populating from HTTP request and response data. Context objects are attached to Transactions or Errors, and will be pooled along with them. Furthermore, space for custom context will be reused, and we will extend this to tags and other fields in the future.

The primary benefits of this are:
 - No more exposure of model types via Transaction or Error, so we can update the model types without affecting the API (#46).
 - Fewer allocations, faster encoding.

benchcmp for contrib/apmhttp

```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkWithoutMiddleware//hello/world-4     600           619           +3.17%
BenchmarkWithoutMiddleware//sleep/1ms-4       1186794       1193345       +0.55%
BenchmarkWithMiddleware//hello/world-4        3722          2925          -21.41%
BenchmarkWithMiddleware//sleep/1ms-4          1217983       1219617       +0.13%

benchmark                                  old allocs     new allocs     delta
BenchmarkWithMiddleware//hello/world-4     15             10             -33.33%
BenchmarkWithMiddleware//sleep/1ms-4       11             6              -45.45%

benchmark                                  old bytes     new bytes     delta
BenchmarkWithMiddleware//hello/world-4     986           605           -38.64%
BenchmarkWithMiddleware//sleep/1ms-4       865           498           -42.43%
```

and for contrib/apmgin:
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkWithoutMiddleware//hello/world-4     1148          1171          +2.00%
BenchmarkWithoutMiddleware//sleep/1ms-4       1138535       1144997       +0.57%
BenchmarkWithMiddleware//hello/world-4        9001          8513          -5.42%
BenchmarkWithMiddleware//sleep/1ms-4          1257260       1215869       -3.29%

benchmark                                     old allocs     new allocs     delta
BenchmarkWithoutMiddleware//hello/world-4     3              3              +0.00%
BenchmarkWithoutMiddleware//sleep/1ms-4       0              0              +0.00%
BenchmarkWithMiddleware//hello/world-4        20             10             -50.00%
BenchmarkWithMiddleware//sleep/1ms-4          16             7              -56.25%

benchmark                                     old bytes     new bytes     delta
BenchmarkWithoutMiddleware//hello/world-4     120           120           +0.00%
BenchmarkWithoutMiddleware//sleep/1ms-4       0             0             +0.00%
BenchmarkWithMiddleware//hello/world-4        1687          639           -62.12%
BenchmarkWithMiddleware//sleep/1ms-4          1538          508           -66.97%
```